requirements.txt
meshtastic>=2.3
pypubsub>=4.0.3